1.1.5 release: Changed Structure block tiers to follow a x/2x linear pattern instead of a 2^x/2^x+1 pattern.  For existing structures to
see this change the structure blocks need to be broken and replaced.
Fixed structure block drop bug.
Cleaned up some duplicate methods and code.  Things should be slightly faster and files slightly smaller now.
Rescaled Industrial Furnace stats so yield scales from ~2x to ~6x, cook/fuel times scales from ~95% to ~25% depending on tier and furnace size.  Changing the tier of the majority of the blocks in the structure should have a more noticeable effect now.
Added in an upper text field for the guidebook
Added in localization support the the guidebook with .lang files
Added in fluid transporter block with a snazzy model. :3
Removed the ability to craft small piles of graphite dust and carbon nanotubes.